backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-mail==0.10.0
pymongo==4.6.2
python-dotenv==1.0.1 